requirements.txt
torch
datasets
transformers
sentencepiece
fasttext
streamlit_option_menu
